+++ config/config_server_develop.yaml
@@ -1,4 +1,4 @@
 iiot_server:
   tcp_server:
-    ip_address: 'localhost'
-    port: 9999
+    ip_address: '192.168.0.241'
+    port: 1233
